+++ src/soopervisor/commons/__init__.py
@@ -1,5 +1,7 @@
 from soopervisor.commons import conda, docker, source, dependencies
-from soopervisor.commons.dag import load_tasks, load_dag, find_spec
+from soopervisor.commons.dag import (load_tasks, load_dag,
+                                     find_spec,
+                                     product_prefixes_from_spec)
 
 __all__ = [
     'conda',
@@ -9,4 +11,5 @@ __all__ = [
     'load_dag',
     'find_spec',
     'dependencies',
+    'product_prefixes_from_spec',
 ]